deploy-package/backend/requirements.txt
Django==6.0.3
djangorestframework==3.17.1
mysqlclient==2.2.8
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.11
python-decouple==3.8
Pillow==12.1.1
django-filter==25.2
gunicorn==23.0.0
